+++ fabric/src/main/java/io/github/sakurawald/fuji/module/mixin/command_attachment/ServerPlayerEntityMixin.java
@@ -8,7 +8,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Hand;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -18,15 +17,13 @@ import java.util.List;
 @Mixin(ServerPlayerEntity.class)
 public abstract class ServerPlayerEntityMixin {
 
-    @Unique
-    final ServerPlayerEntity player = (ServerPlayerEntity) (Object) this;
-
     @Inject(method = "swingHand", at = @At("HEAD"))
     void onPlayerLeftClick(Hand hand, CallbackInfo ci) {
+        ServerPlayerEntity player = (ServerPlayerEntity) (Object) this;
+
         if (hand.equals(Hand.MAIN_HAND)) {
             ItemStack mainHandStack = player.getMainHandStack();
             String uuid = UuidHelper.getAttachedUuid(ItemStackHelper.CustomData.getCustomDataNbt(mainHandStack));
-            ServerPlayerEntity player = (ServerPlayerEntity) (Object) this;
 
             CommandAttachmentService.tryTriggerAttachmentDataNode(uuid, player, List.of(InteractType.LEFT_CLICK, InteractType.ANY_CLICK), () -> {});
         }