+++ CHANGELOG.md
@@ -43,3 +43,4 @@
   - feature: improve the `readbility` of `/command-attach query-{block|entity|item}` commands.
   - feature: support `SWAP_HAND` interaction type.
   - docs: new documents for this module.
+  - feature: prevent the `block` being broken, if the there are `command attachments` attached in this block.